+++ openmetadata-ui/src/main/resources/ui/src/components/DataModels/DataModelsTable.tsx
@@ -13,27 +13,41 @@
 
 import { Col } from 'antd';
 import { ColumnsType } from 'antd/lib/table';
+import { AxiosError } from 'axios';
 import { isUndefined } from 'lodash';
-import React, { useMemo } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Link } from 'react-router-dom';
+import { Link, useParams } from 'react-router-dom';
 import ErrorPlaceHolder from '../../components/common/ErrorWithPlaceholder/ErrorPlaceHolder';
 import RichTextEditorPreviewer from '../../components/common/RichTextEditor/RichTextEditorPreviewer';
 import Table from '../../components/common/Table/Table';
-import { getDataModelDetailsPath, PAGE_SIZE } from '../../constants/constants';
-import { DataModelTableProps } from '../../pages/DataModelPage/DataModelsInterface';
+import {
+  getDataModelDetailsPath,
+  pagingObject,
+} from '../../constants/constants';
+import { Paging } from '../../generated/type/paging';
+import { usePaging } from '../../hooks/paging/usePaging';
 import { ServicePageData } from '../../pages/ServiceDetailsPage/ServiceDetailsPage';
+import { getDataModels } from '../../rest/dashboardAPI';
 import { getEntityName } from '../../utils/EntityUtils';
+import { showPagination } from '../../utils/Pagination/PaginationUtils';
+import { showErrorToast } from '../../utils/ToastUtils';
 import NextPrevious from '../common/NextPrevious/NextPrevious';
+import { NextPreviousProps } from '../common/NextPrevious/NextPrevious.interface';
 
-const DataModelTable = ({
-  data,
-  isLoading,
-  paging,
-  pagingHandler,
-  currentPage,
-}: DataModelTableProps) => {
+const DataModelTable = () => {
   const { t } = useTranslation();
+  const { fqn } = useParams<{ fqn: string }>();
+  const [dataModels, setDataModels] = useState<Array<ServicePageData>>();
+  const {
+    currentPage,
+    pageSize,
+    paging,
+    handlePageChange,
+    handlePageSizeChange,
+    handlePagingChange,
+  } = usePaging();
+  const [isLoading, setIsLoading] = useState(true);
 
   const tableColumn: ColumnsType<ServicePageData> = useMemo(
     () => [
@@ -73,6 +87,43 @@ const DataModelTable = ({
     []
   );
 
+  const fetchDashboardsDataModel = useCallback(
+    async (pagingData?: Partial<Paging>) => {
+      try {
+        setIsLoading(true);
+        const { data, paging: resPaging } = await getDataModels({
+          service: fqn,
+          fields: 'owner,tags,followers',
+          limit: pageSize,
+          ...pagingData,
+        });
+        setDataModels(data);
+        handlePagingChange(resPaging);
+      } catch (error) {
+        showErrorToast(error as AxiosError);
+        setDataModels([]);
+        handlePagingChange(pagingObject);
+      } finally {
+        setIsLoading(false);
+      }
+    },
+    [fqn, pageSize]
+  );
+
+  const handleDataModelPageChange: NextPreviousProps['pagingHandler'] = ({
+    cursorType,
+    currentPage,
+  }) => {
+    if (cursorType) {
+      fetchDashboardsDataModel({ [cursorType]: paging[cursorType] });
+    }
+    handlePageChange(currentPage);
+  };
+
+  useEffect(() => {
+    fetchDashboardsDataModel();
+  }, [pageSize]);
+
   return (
     <>
       <Col className="p-x-lg" data-testid="table-container" span={24}>
@@ -81,7 +132,7 @@ const DataModelTable = ({
           className="mt-4 table-shadow"
           columns={tableColumn}
           data-testid="data-models-table"
-          dataSource={data}
+          dataSource={dataModels}
           loading={isLoading}
           locale={{
             emptyText: <ErrorPlaceHolder className="m-y-md" />,
@@ -92,12 +143,13 @@ const DataModelTable = ({
         />
       </Col>
       <Col span={24}>
-        {paging && paging.total > PAGE_SIZE && (
+        {showPagination(paging, pageSize) && (
           <NextPrevious
             currentPage={currentPage}
-            pageSize={PAGE_SIZE}
+            pageSize={pageSize}
             paging={paging}
-            pagingHandler={pagingHandler}
+            pagingHandler={handleDataModelPageChange}
+            onShowSizeChange={handlePageSizeChange}
           />
         )}
       </Col>
