+++ openmetadata-ui/src/main/resources/ui/src/components/TableQueries/TableQueries.tsx
@@ -15,7 +15,6 @@ import { Button, Col, Row, Space, Tooltip, Typography } from 'antd';
 import { AxiosError } from 'axios';
 import { compare } from 'fast-json-patch';
 import { isUndefined } from 'lodash';
-import { PagingResponse } from 'Models';
 import Qs from 'qs';
 import React, { FC, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -25,11 +24,7 @@ import {
   OperationPermission,
   ResourceEntity,
 } from '../../components/PermissionProvider/PermissionProvider.interface';
-import {
-  INITIAL_PAGING_VALUE,
-  PAGE_SIZE,
-  pagingObject,
-} from '../../constants/constants';
+import { PAGE_SIZE } from '../../constants/constants';
 import { USAGE_DOCS } from '../../constants/docs.constants';
 import { NO_PERMISSION_FOR_ACTION } from '../../constants/HelperTextUtil';
 import {
@@ -38,6 +33,7 @@ import {
 } from '../../constants/Query.constant';
 import { ERROR_PLACEHOLDER_TYPE } from '../../enums/common.enum';
 import { Query } from '../../generated/entity/data/query';
+import { usePaging } from '../../hooks/paging/usePaging';
 import {
   getQueriesList,
   getQueryById,
@@ -45,6 +41,7 @@ import {
   patchQueries,
   updateQueryVote,
 } from '../../rest/queryAPI';
+import { showPagination } from '../../utils/Pagination/PaginationUtils';
 import { DEFAULT_ENTITY_PERMISSION } from '../../utils/PermissionsUtils';
 import {
   parseSearchParams,
@@ -75,19 +72,22 @@ const TableQueries: FC<TableQueriesProp> = ({
     return searchData;
   }, [location]);
 
-  const [tableQueries, setTableQueries] = useState<PagingResponse<Query[]>>({
-    data: [],
-    paging: pagingObject,
-  });
+  const [tableQueries, setTableQueries] = useState<Query[]>([]);
   const [isLoading, setIsLoading] = useState(QUERY_PAGE_LOADING_STATE);
   const [isError, setIsError] = useState(QUERY_PAGE_ERROR_STATE);
   const [selectedQuery, setSelectedQuery] = useState<Query>();
   const [queryPermissions, setQueryPermissions] = useState<OperationPermission>(
     DEFAULT_ENTITY_PERMISSION
   );
-  const [currentPage, setCurrentPage] = useState(
-    Number(searchParams.queryFrom) || INITIAL_PAGING_VALUE
-  );
+
+  const {
+    currentPage,
+    handlePageChange,
+    pageSize,
+    handlePageSizeChange,
+    paging,
+    handlePagingChange,
+  } = usePaging();
 
   const { getEntityPermission, permissions } = usePermissionProvider();
 
@@ -133,7 +133,7 @@ const TableQueries: FC<TableQueriesProp> = ({
       setTableQueries((pre) => {
         return {
           ...pre,
-          data: pre.data.map((query) =>
+          data: pre.map((query) =>
             query.id === updatedQuery.id ? { ...query, ...res } : query
           ),
         };
@@ -153,7 +153,7 @@ const TableQueries: FC<TableQueriesProp> = ({
       setTableQueries((pre) => {
         return {
           ...pre,
-          data: pre.data.map((query) =>
+          data: pre.map((query) =>
             query.id === response.id ? response : query
           ),
         };
@@ -168,22 +168,22 @@ const TableQueries: FC<TableQueriesProp> = ({
   ) => {
     setIsLoading((pre) => ({ ...pre, query: true }));
     try {
-      const queries = await getQueriesList({
+      const { data: queries, paging } = await getQueriesList({
         ...params,
         limit: PAGE_SIZE,
         entityId: tableId,
         fields: 'owner,votes,tags,queryUsedIn,users',
       });
-      if (queries.data.length === 0) {
+      if (queries.length === 0) {
         setIsError((pre) => ({ ...pre, page: true }));
       } else {
         setTableQueries(queries);
         const selectedQueryData = searchParams.query
-          ? queries.data.find((query) => query.id === searchParams.query) ||
-            queries.data[0]
-          : queries.data[0];
+          ? queries.find((query) => query.id === searchParams.query) ||
+            queries[0]
+          : queries[0];
         setSelectedQuery(selectedQueryData);
-
+        handlePagingChange(paging);
         history.push({
           search: stringifySearchParams({
             tableId,
@@ -202,10 +202,9 @@ const TableQueries: FC<TableQueriesProp> = ({
   };
 
   const pagingHandler = ({ cursorType, currentPage }: PagingHandlerParams) => {
-    const { paging } = tableQueries;
     if (cursorType) {
       fetchTableQuery({ [cursorType]: paging[cursorType] }, currentPage);
-      setCurrentPage(currentPage);
+      handlePageChange(currentPage);
     }
   };
 
@@ -283,7 +282,7 @@ const TableQueries: FC<TableQueriesProp> = ({
       </ErrorPlaceHolder>
     </Col>
   ) : (
-    tableQueries.data.map((query) => (
+    tableQueries.map((query) => (
       <Col data-testid="query-card" key={query.id} span={24}>
         <QueryCard
           afterDeleteAction={fetchTableQuery}
@@ -313,12 +312,13 @@ const TableQueries: FC<TableQueriesProp> = ({
           {isLoading.query ? <Loader /> : queryTabBody}
 
           <Col span={24}>
-            {tableQueries.paging.total > PAGE_SIZE && (
+            {showPagination(paging, pageSize) && (
               <NextPrevious
                 currentPage={currentPage}
-                pageSize={PAGE_SIZE}
-                paging={tableQueries.paging}
+                pageSize={pageSize}
+                paging={paging}
                 pagingHandler={pagingHandler}
+                onShowSizeChange={handlePageSizeChange}
               />
             )}
           </Col>
