+++ openmetadata-ui/src/main/resources/ui/src/components/Team/TeamDetails/UserTab/UserTab.component.tsx
@@ -15,7 +15,7 @@ import { Button, Col, Row, Space, Tooltip } from 'antd';
 import { ColumnsType } from 'antd/lib/table';
 import { isEmpty, orderBy } from 'lodash';
 import QueryString from 'qs';
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { ReactComponent as ExportIcon } from '../../../../assets/svg/ic-export.svg';
@@ -28,36 +28,43 @@ import Table from '../../../../components/common/Table/Table';
 import { UserSelectableList } from '../../../../components/common/UserSelectableList/UserSelectableList.component';
 import { useEntityExportModalProvider } from '../../../../components/Entity/EntityExportModalProvider/EntityExportModalProvider.component';
 import ConfirmationModal from '../../../../components/Modals/ConfirmationModal/ConfirmationModal';
-import { PAGE_SIZE_MEDIUM } from '../../../../constants/constants';
+import {
+  INITIAL_PAGING_VALUE,
+  PAGE_SIZE_BASE,
+  PAGE_SIZE_MEDIUM,
+} from '../../../../constants/constants';
 import {
   GlobalSettingOptions,
   GlobalSettingsMenuCategory,
 } from '../../../../constants/GlobalSettings.constants';
 import { ERROR_PLACEHOLDER_TYPE } from '../../../../enums/common.enum';
 import { EntityAction } from '../../../../enums/entity.enum';
+import { SearchIndex } from '../../../../enums/search.enum';
 import { User } from '../../../../generated/entity/teams/user';
 import { EntityReference } from '../../../../generated/entity/type';
+import { Paging } from '../../../../generated/type/paging';
+import { usePaging } from '../../../../hooks/paging/usePaging';
+import { SearchResponse } from '../../../../interface/search.interface';
 import { ImportType } from '../../../../pages/TeamsPage/ImportTeamsPage/ImportTeamsPage.interface';
+import { searchData } from '../../../../rest/miscAPI';
 import { exportUserOfTeam } from '../../../../rest/teamsAPI';
+import { getUsers } from '../../../../rest/userAPI';
+import { formatUsersResponse } from '../../../../utils/APIUtils';
 import { getEntityName } from '../../../../utils/EntityUtils';
+import { showPagination } from '../../../../utils/Pagination/PaginationUtils';
 import { getSettingsPathWithFqn } from '../../../../utils/RouterUtils';
+import { getDecodedFqn, getEncodedFqn } from '../../../../utils/StringsUtils';
 import { commonUserDetailColumns } from '../../../../utils/Users.util';
 import ManageButton from '../../../common/EntityPageInfos/ManageButton/ManageButton';
 import NextPrevious from '../../../common/NextPrevious/NextPrevious';
+import { PagingHandlerParams } from '../../../common/NextPrevious/NextPrevious.interface';
 import Searchbar from '../../../common/SearchBarComponent/SearchBar.component';
 import { UserTabProps } from './UserTab.interface';
 
 export const UserTab = ({
-  users,
-  searchText,
-  isLoading,
   permission,
   currentTeam,
-  onSearchUsers,
   onAddUser,
-  paging,
-  onChangePaging,
-  currentPage,
   onRemoveUser,
 }: UserTabProps) => {
   const { t } = useTranslation();
@@ -69,6 +76,102 @@ export const UserTab = ({
     const user = currentTeam.users?.find((u) => u.id === id);
     setDeletingUser(user);
   };
+  const [isLoading, setIsLoading] = useState(true);
+  const [users, setUsers] = useState<User[]>([]);
+  const [searchText, setSearchText] = useState('');
+  const {
+    currentPage,
+    pageSize,
+    paging,
+    handlePageChange,
+    handlePageSizeChange,
+    handlePagingChange,
+  } = usePaging(PAGE_SIZE_MEDIUM);
+
+  /**
+   * Make API call to fetch current team user data
+   */
+  const getCurrentTeamUsers = (team: string, paging: Partial<Paging> = {}) => {
+    setIsLoading(true);
+    getUsers({
+      fields: 'teams,roles',
+      limit: PAGE_SIZE_BASE,
+      team: getDecodedFqn(team),
+      ...paging,
+    })
+      .then((res) => {
+        if (res.data) {
+          setUsers(res.data);
+          handlePagingChange(res.paging);
+        }
+      })
+      .catch(() => {
+        setUsers([]);
+        handlePagingChange({ total: 0 });
+      })
+      .finally(() => {
+        setIsLoading(false);
+      });
+  };
+
+  const searchUsers = (text: string, currentPage: number) => {
+    setIsLoading(true);
+    searchData(
+      text,
+      currentPage,
+      PAGE_SIZE_BASE,
+      `(teams.id:${currentTeam?.id})`,
+      '',
+      '',
+      SearchIndex.USER
+    )
+      .then((res) => {
+        const data = formatUsersResponse(
+          (res.data as SearchResponse<SearchIndex.USER>).hits.hits
+        );
+        setUsers(data);
+        handlePagingChange({
+          total: res.data.hits.total.value,
+        });
+      })
+      .catch(() => {
+        setUsers([]);
+      })
+      .finally(() => setIsLoading(false));
+  };
+
+  const userPagingHandler = ({
+    cursorType,
+    currentPage,
+  }: PagingHandlerParams) => {
+    if (searchText) {
+      handlePageChange(currentPage);
+      searchUsers(searchText, currentPage);
+    } else if (cursorType) {
+      handlePageChange(currentPage);
+      getCurrentTeamUsers(currentTeam.name, {
+        [cursorType]: paging[cursorType],
+      });
+    }
+  };
+
+  const handleCurrentUserPage = (value?: number) => {
+    handlePageChange(value ?? INITIAL_PAGING_VALUE);
+  };
+
+  const handleUsersSearchAction = (text: string) => {
+    setSearchText(text);
+    handleCurrentUserPage(INITIAL_PAGING_VALUE);
+    if (text) {
+      searchUsers(text, INITIAL_PAGING_VALUE);
+    } else {
+      getCurrentTeamUsers(currentTeam.name);
+    }
+  };
+
+  useEffect(() => {
+    getCurrentTeamUsers(getEncodedFqn(currentTeam.name));
+  }, [currentTeam]);
 
   const columns: ColumnsType<User> = useMemo(() => {
     return [
@@ -175,7 +278,7 @@ export const UserTab = ({
     }
   };
 
-  if (isEmpty(users) && !searchText && isLoading <= 0) {
+  if (isEmpty(users) && !searchText) {
     return (
       <ErrorPlaceHolder
         button={
@@ -225,7 +328,7 @@ export const UserTab = ({
               })}
               searchValue={searchText}
               typingInterval={500}
-              onSearch={onSearchUsers}
+              onSearch={handleUsersSearchAction}
             />
           </Col>
           <Col>
@@ -255,7 +358,7 @@ export const UserTab = ({
           className="teams-list-table"
           columns={columns}
           dataSource={sortedUser}
-          loading={isLoading > 0}
+          loading={isLoading}
           locale={{
             emptyText: <FilterTablePlaceHolder />,
           }}
@@ -263,13 +366,14 @@ export const UserTab = ({
           rowKey="name"
           size="small"
         />
-        {paging.total > PAGE_SIZE_MEDIUM && (
+        {showPagination(paging, pageSize) && (
           <NextPrevious
             currentPage={currentPage}
             isNumberBased={Boolean(searchText)}
-            pageSize={PAGE_SIZE_MEDIUM}
+            pageSize={pageSize}
             paging={paging}
-            pagingHandler={onChangePaging}
+            pagingHandler={userPagingHandler}
+            onShowSizeChange={handlePageSizeChange}
           />
         )}
       </Col>
